pyproject.toml
[project]
name = "syntaclip"
version = "0.1.0"
description = "This is an implementation of CLIP that takes Syntactic Distance into account."
authors = [
    { name = "n4okins", email = "63445867+n4okins@users.noreply.github.com" }
]
dependencies = [
    "torch==2.4.0+cu121",
    "torchvision==0.19.0+cu121",
    "torchinfo>=1.8.0",
    "transformers>=4.43.3",
    "matplotlib>=3.9.2",
    "nvidia-dali-cuda120>=1.40.0",
    "nvidia-nvjpeg2k-cu12>=0.8.0.38"
]
readme = "README.md"
requires-python = ">= 3.10"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = [
    "jupyterlab>=4",
    "hydra-core>=1.3.2",
    "wandb>=0.16.0",
    "pytest>=8.3.2",
    "ruff>=0.6.2",
    "mypy>=1.11.2",
    "utils @ git+https://github.com/n4okins/utils.git",
    "img2dataset>=1.45.0",
    "ipywidgets>=8.1.3",
    "open-clip-torch>=2.26.1",
]

[[tool.rye.sources]]
name = "pytorch"
url = "https://download.pytorch.org/whl/cu121"
type = "index"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/syntaclip"]

[tool.mypy]
exclude = [
    "ignores/"
]
ignore_missing_imports = true
disable_error_code = "import-untyped"
